refactor: add Meziantou and Roslynator analyzers, fix core library issues
Added code quality analyzers:
- Meziantou.Analyzer 3.0.44
- Roslynator.Analyzers 4.12.4

Fixed ZeroAlloc.EventSourcing core library (all issues resolved):
- MA0004: Added ConfigureAwait(false) to 26+ library awaits
- MA0048: Extracted type-named files (CommitStrategy.cs, ErrorHandlingStrategy.cs, ProjectionConsistencyReport.cs)
- MA0025: Changed NotImplementedException to NotSupportedException for dead-letter
- MA0026: Removed TODO comment
- MA0002: Added StringComparer.Ordinal to ConcurrentDictionary

Result: Core library builds with 0 errors, 0 warnings

Remaining: Other projects (PostgreSQL, Aggregates, Kafka, SQL) have method length and additional ConfigureAwait issues to address in follow-up.
