+++ Directory.Packages.props
@@ -10,6 +10,8 @@
     <PackageVersion Include="ZeroAlloc.Serialisation"       Version="1.0.0" />
     <PackageVersion Include="ZeroAlloc.Pipeline"            Version="0.1.5" />
     <PackageVersion Include="ZeroAlloc.Analyzers"           Version="1.3.12" />
+    <PackageVersion Include="Meziantou.Analyzer"            Version="3.0.44" />
+    <PackageVersion Include="Roslynator.Analyzers"          Version="4.12.4" />
     <PackageVersion Include="xunit"                         Version="2.9.3" />
     <PackageVersion Include="xunit.runner.visualstudio"     Version="3.1.5" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk"        Version="18.4.0" />
