+++ src/ZeroAlloc.EventSourcing/ZeroAlloc.EventSourcing.csproj
@@ -11,4 +11,9 @@
     <PackageReference Include="ZeroAlloc.ValueObjects" />
   </ItemGroup>
 
+  <ItemGroup Condition="'$(IsRoslynComponent)' != 'true'">
+    <PackageReference Include="Meziantou.Analyzer" PrivateAssets="all" />
+    <PackageReference Include="Roslynator.Analyzers" PrivateAssets="all" />
+  </ItemGroup>
+
 </Project>
